Reorganize data_scrapers and news into one package
This way, the two can share code. I've changed as little code as possible here -- mostly just made a lot of absolute imports into relative imports. The two top-level scripts are still the same so the frontend does have to change how it uses this repo at all. The `data_models` directory is still at the top level since it's not really Python code.

The top-level package is now named `covid19_sfbayarea`, which is based on the git repo's name.
