+++ covid19_sfbayarea/news/base.py
