+++ covid19_sfbayarea/news/santa_clara.py
