+++ scraper_news.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import click
 import json
-import news
+from covid19_sfbayarea import news
 from pathlib import Path
 from typing import Tuple
